add rpc listener
