+++ devcontainer/main.go
@@ -17,59 +17,14 @@ import (
 
 type args []string
 
-func (a args) Add(b args) {
-	a = append(a, b...)
+func (a *args) Add(b args) {
+	*a = append(*a, b...)
+}
+func (a *args) AddString(s string) {
+	a.Add(args{s})
 }
 func (a args) String() string {
 	return strings.Join(a, " ")
-}
-
-type object map[string]any
-type stringOrArray any
-type stringOrArrayOrInt any
-
-// refer to https://containers.dev/implementors/json_reference/
-type cfgType struct {
-	Image string `json:"image"`
-	Build struct {
-		Dockerfile string        `json:"dockerfile"`
-		Context    string        `json:"context"`
-		Args       object        `json:"args"`
-		Target     string        `json:"target"`
-		CacheFrom  stringOrArray `json:"cacheFrom"`
-	} `json:"build"`
-	Settings        object             `json:"settings"`
-	AppPort         stringOrArrayOrInt `json:"appPort"`
-	ContainerEnv    object             `json:"containerEnv"`
-	ContainerUser   string             `json:"containerUser"`
-	Mounts          []string           `json:"mounts"`
-	WorkspaceMount  string             `json:"workspaceMount"`
-	WorkspaceFolder string             `json:"workspaceFolder"`
-	RunArgs         []string           `json:"runArgs"`
-
-	DockerComposeFile stringOrArray `json:"dockerComposeFile"`
-	Service           string        `json:"service"`
-	RunServices       []string      `json:"runServices"`
-
-	Name                 string        `json:"name"`
-	ForwardPorts         stringOrArray `json:"forwardPorts"`
-	PortsAttributes      object        `json:"portsAttributes"`
-	OtherPortsAttributes object        `json:"otherPortsAttributes"`
-	RemoteEnv            object        `json:"remoteEnv"`
-	RemoteUser           string        `json:"remoteUser"`
-	UpdateRemoteUserUID  bool          `json:"updateRemoteUserUID"`
-	UserEnvProbe         string        `json:"userEnvProbe"`
-	OverrideCommand      bool          `json:"overrideCommand"`
-	Features             object        `json:"features"`
-	ShutdownAction       string        `json:"shutdownAction"`
-	Customizations       object        `json:"customizations"`
-	InitializeCommand    stringOrArray `json:"initializeCommand"`
-	OnCreateCommand      stringOrArray `json:"onCreateCommand"`
-	UpdateContentCommand stringOrArray `json:"updateContentCommand"`
-	PostCreateCommand    stringOrArray `json:"postCreateCommand"`
-	PostStartCommand     stringOrArray `json:"postStartCommand"`
-	PostAttachCommand    stringOrArray `json:"postAttachCommand"`
-	WaitFor              string        `json:"waitFor"`
 }
 
 func formatPorts(v any) []string {
@@ -107,20 +62,18 @@ func formatMount(src, dst string) []string {
 	return []string{"--mount", fmt.Sprintf("type=bind,source=%s,target=%s", src, dst)}
 }
 
-func getWd() string {
-	if d, err := os.Getwd(); err == nil {
-		return d
-	}
-	return ""
-}
-
 func main() {
-	// TODO: add local socket here that can listen for commands?
-	dir := flag.String("dir", ".devcontainer", "directory with devcontainer.json and Dockerfile")
-	wd := flag.String("l", getWd(), "local workspace to map into container")
+	sock := flag.String("s", "/tmp/devcontainer.sock", "unix socket to listen on for rpc commands")
+	dir := flag.String("d", ".devcontainer", "directory with devcontainer.json and Dockerfile")
+	wd := flag.String("l", ".", "local workspace to map into container")
 	ws := flag.String("w", "/workspace", "Working directory inside the container")
-	cmd := flag.String("cmd", "docker", "name of docker command")
+	cmd := flag.String("docker", "docker", "name of docker command")
 	flag.Parse()
+	if len(flag.Args()) > 0 {
+		payload := strings.Join(flag.Args(), " ")
+		sendRpc(*sock, payload)
+		return
+	}
 	containerFile := filepath.Join(*dir, "devcontainer.json")
 	dockerFile := filepath.Join(*dir, "Dockerfile")
 	log.Println("reading file", containerFile)
@@ -142,7 +95,10 @@ func main() {
 	if err := decoder.Decode(&cfg); err != nil {
 		log.Fatal("error decoding file", err)
 	}
-	log.Println("config:", cfg)
+	ctrl := new(Control)
+	rpcListener(*sock, ctrl)
+	defer os.Remove(*sock)
+	//log.Println("config:", cfg)
 	buildTag := fmt.Sprintf("localdevcon-%s:%d", strings.ToLower(cfg.Name), time.Now().Unix())
 	buildArgs := args{"build", "-t", buildTag, "-f", dockerFile}
 	buildArgs.Add(formatObject("--build-arg", cfg.Build.Args))
